+++ src/net/fexcraft/app/fmt/ui/tree/ModelTree.java
@@ -67,7 +67,7 @@ public class ModelTree extends RightTree {
 	}
 
 	@Override
-	public void hovered(int mx, int my){
+	public void hovered(float mx, float my){
 		super.hovered(mx, my);
 	}
